requirements.txt
requests==2.32.3
snowflake-connector-python==3.13.1
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.67.1
beautifulsoup4==4.12.3
tabulate==0.9.0
